// src/backend/src/properties/dto/property-filter.dto.ts
export class PropertyFilterDto {
  q?: string;
  type?: string;
  status?: string;
  city?: string;
  district?: string;
  ward?: string;
  regionId?: string;
  minPrice?: string;
  maxPrice?: string;
  minArea?: string;
  maxArea?: string;
  minBedroom?: string;
  maxBedroom?: string;
  minBathroom?: string;
  maxBathroom?: string;
  minFloor?: string;
  maxFloor?: string;
  includeHidden?: string;
  sortBy?: string;
  sortOrder?: string;
  page?: string;
  limit?: string;
}